requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asteval==0.9.31
boto3==1.34.26
botocore==1.34.26
certifi==2023.11.17
charset-normalizer==3.3.2
distro==1.8.0
dnspython==2.4.2
exceptiongroup==1.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
jmespath==1.0.1
openai==1.3.4
Pillow==10.1.0
pydantic==2.5.2
pydantic_core==2.14.5
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7